+++ test/tunicode.c
@@ -807,7 +807,7 @@ void test_unicode(void)
   MESSAGE(5, ("Testing UTF-8 Encoding\n"));
 
   /* Create a random string with length NUM_CHARS */
-  HDsrandom((unsigned long)HDtime(NULL));
+  HDsrandom((unsigned)HDtime(NULL));
 
   HDmemset(test_string, 0, sizeof(test_string));
   for(x=0; x<NUM_CHARS; x++)
